requirements.txt
addict==2.4.0
numpy==1.22.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
chumpy==0.70
click==8.1.3
cycler==0.11.0
Cython==0.29.30
decorator==4.4.2
Flask==2.1.2
fonttools==4.33.3
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
json-tricks==3.15.5
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
#mmcv-full==1.4.7 -f https://download.openmmlab.com/mmcv/dist/cu102/torch1.11.0/index.html
#-e git+https://github.com/open-mmlab/mmdetection.git@ca11860f4f3c3ca2ce8340e2686eeaec05b29111#egg=mmdet
mmpose==0.22.0
moviepy==1.0.3
munkres==1.1.4
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
proglog==0.1.10
pycocotools
pyparsing==3.0.9
python-dateutil==2.8.2
pytorchcv==0.0.67
PyYAML==6.0
requests==2.27.1
scipy==1.8.1
six==1.16.0
terminaltables==3.1.10
#torch==1.11.0
#torchaudio==0.11.0
#torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
xtcocotools==1.12
yapf==0.32.0
zipp==3.8.0
